+++ react/modules/imageAsList.tsx
@@ -8,7 +8,6 @@ export const getImagesAsJSXList = (
   isMobile: boolean,
   height: string | number,
   preload?: boolean,
-  loading?: 'eager' | 'lazy' | undefined,
   experimentalPreventLayoutShift?: boolean,
   experimentalSetExplicitDimensions?: boolean
 ) => {
@@ -39,7 +38,6 @@ export const getImagesAsJSXList = (
           experimentalSetExplicitDimensionsChild
         }
         preload={preload && idx === 0}
-        loading={loading}
         {...props}
       />
     )